Alias `quit_driver` to `driver_quit`.
To start a driver, you create a driver object then call `start_driver`.

To quit a driver, you have to remember both not to call `stop_driver`, but also
to reverse the order and call `driver_quit`, which just feels a bit... janky?

I'd suggest also aliasing `stop_driver`, but calling `quit` to end a controlled
session seems to be the community norm.